StatusLights: break up code into multiple files
