+++ applications/services/status_lights/status_lights.c
@@ -1,54 +1,10 @@
-#include "status_lights.h"
-#include "status_lights_common_private.h"
-
-#include <intercom/intercom.h>
-
-#include <api_lock.h>
+#include "status_lights_i.h"
 
 #define TAG "StatusLights"
 
-#define STATUS_LIGHTS_BRIGHTNESS_MIN     (0)
-#define STATUS_LIGHTS_BRIGHTNESS_MAX     (100)
-#define STATUS_LIGHTS_BRIGHTNESS_DEFAULT (50)
+typedef void (*StatusLightsApiMessageHandler)(StatusLights* instance, StatusLightsMessage* message);
 
-struct StatusLights {
-    FuriEventLoop* event_loop;
-    FuriMessageQueue* message_queue;
-    IntercomChannel* intercom_ch;
-
-    StatusLightsBrightness brightness;
-};
-
-typedef enum {
-    StatusLightsMessageTypeSetBrightness,
-    StatusLightsMessageTypeGetBrightness,
-    StatusLightsMessageTypeSetRunPreset,
-
-    StatusLightsMessageTypesCount,
-} StatusLightsMessageType;
-
-typedef struct {
-    FuriApiLock api_lock;
-    StatusLightsMessageType type;
-    union {
-        struct {
-            StatusLightsBrightness brightness;
-        } as_set_brightness;
-
-        struct {
-            StatusLightsBrightness* brightness;
-        } as_get_brightness;
-
-        struct {
-            StatusLightsPreset preset;
-            Color color;
-        } as_run_preset;
-    };
-} StatusLightsMessage;
-
-typedef void (*MessageHandler)(StatusLights* instance, StatusLightsMessage* message);
-
-static const MessageHandler message_handlers[];
+static const StatusLightsApiMessageHandler api_message_handlers[];
 
 static void status_lights_send_command(StatusLights* instance, StatusLightsCommand* command) {
     size_t tx_size =
@@ -104,7 +60,7 @@ static void status_lights_message_callback(FuriEventLoopObject* object, void* co
         furi_message_queue_get(instance->message_queue, &message, FuriWaitForever) ==
         FuriStatusOk);
 
-    message_handlers[message.type](instance, &message);
+    api_message_handlers[message.type](instance, &message);
 
     if(message.api_lock) {
         api_lock_unlock(message.api_lock);
@@ -143,68 +99,10 @@ int32_t status_lights_srv(void* p) {
     return 0;
 }
 
-void status_lights_run_preset(StatusLights* instance, StatusLightsPreset preset, Color color) {
-    furi_check(instance);
-    furi_check(preset < StatusLightsPresetsCount);
-
-    StatusLightsMessage message = {
-        .api_lock = NULL,
-        .type = StatusLightsMessageTypeSetRunPreset,
-        .as_run_preset =
-            {
-                .preset = preset,
-                .color = color,
-            },
-    };
-
-    furi_check(
-        furi_message_queue_put(instance->message_queue, &message, FuriWaitForever) ==
-        FuriStatusOk);
-}
-
-void status_lights_set_brightness(StatusLights* instance, StatusLightsBrightness brightness) {
-    furi_check(instance);
-
-    StatusLightsMessage message = {
-        .api_lock = NULL,
-        .type = StatusLightsMessageTypeSetBrightness,
-        .as_set_brightness =
-            {
-                .brightness = brightness,
-            },
-    };
-
-    furi_check(
-        furi_message_queue_put(instance->message_queue, &message, FuriWaitForever) ==
-        FuriStatusOk);
-}
-
-StatusLightsBrightness status_lights_get_brightness(StatusLights* instance) {
-    furi_check(instance);
-
-    StatusLightsBrightness brightness;
-    StatusLightsMessage message = {
-        .api_lock = api_lock_alloc_locked(),
-        .type = StatusLightsMessageTypeGetBrightness,
-        .as_get_brightness =
-            {
-                .brightness = &brightness,
-            },
-    };
-
-    furi_check(
-        furi_message_queue_put(instance->message_queue, &message, FuriWaitForever) ==
-        FuriStatusOk);
-
-    api_lock_wait_unlock_and_free(message.api_lock);
-
-    return brightness;
-}
-
-static const MessageHandler message_handlers[] = {
+static const StatusLightsApiMessageHandler api_message_handlers[] = {
     [StatusLightsMessageTypeSetBrightness] = status_lights_do_set_brightness,
     [StatusLightsMessageTypeGetBrightness] = status_lights_do_get_brightness,
     [StatusLightsMessageTypeSetRunPreset] = status_lights_do_run_preset,
 };
 
-static_assert(COUNT_OF(message_handlers) == StatusLightsMessageTypesCount);
+static_assert(COUNT_OF(api_message_handlers) == StatusLightsMessageTypesCount);
